yingshiyun_merged/requirements.txt
# 影视云统一服务依赖

# 基础Web框架
fastapi==0.116.0
uvicorn==0.35.0
starlette>=0.40,<1
pydantic==2.11.7

# HTTP客户端
aiohttp==3.11.18
requests==2.32.3
anyio>=4.0,<5

# 数据库相关
aiomysql==0.2.0
PyMySQL==1.1.0
pymysql>=1.1.0

# AI/LLM相关
openai>=1.51.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain_core==0.3.68
langchain_community==0.3.27
langchain_openai==0.3.27
tiktoken==0.9.0

# 数据科学相关
pandas==2.3.1
numpy==2.2.5
openpyxl>=3.1,<4

# 农历/日期处理
chinese_calendar==1.10.0
lunardate==0.2.2
zhdate==0.1

# 工具库
python-dotenv==1.1.1
pyyaml>=6.0
jinja2==3.1.3

# Redis
redis==5.2.1

# 监控
prometheus_client==0.22.0
psutil==5.9.6
watchdog>=3.0.0

# 安全
cryptography>=41.0.0,<42.0.0

# 编译依赖
cython==3.1.2
setuptools==75.8.0

# 其他
idna==3.6
urllib3==2.4.0
